requirements.txt
# Django Core
Django==5.0.7
python-decouple==3.8

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# API Development
djangorestframework==3.15.2
django-cors-headers==4.4.0

# Media Files
Pillow==10.4.0

# Production Server
gunicorn==22.0.0
whitenoise==6.7.0

# Development Tools
django-debug-toolbar==4.4.6

# Testing
pytest-django==4.8.0
factory-boy==3.3.0

# PostgreSQL Database Adapter
psycopg2-binary==2.9.9
